+++ react/components/BindingSelectorList.tsx
@@ -1,50 +1,56 @@
+import React, { useState } from 'react'
 import type { FC } from 'react'
-import React from 'react'
+import { Button, ButtonGroup } from 'vtex.styleguide'
 import { useCssHandles } from 'vtex.css-handles'
 
+import LabelOption from './LabelsFlags'
+
+const CSS_HANDLES = ['listContainer'] as const
+
 interface Props {
-  open: boolean
-  currentBinding: TranslationsAndSettings
-  bindingInfo: TranslationsAndSettings[]
+  bindingList: TranslationsAndSettings[]
   onSelectBinding: (selectedBinding: TranslationsAndSettings) => void
+  currentBinding: TranslationsAndSettings
+  display: FlagDisplay
+  isLoading: boolean
 }
 
-const CSS_HANDLES = ['list', 'listElement'] as const
-
 const BindingSelectorList: FC<Props> = ({
-  open,
-  currentBinding,
-  bindingInfo,
+  bindingList,
   onSelectBinding,
+  currentBinding,
+  display,
+  isLoading,
 }) => {
+  const [loading, setLoading] = useState(false)
   const handles = useCssHandles(CSS_HANDLES)
 
+  const mappedBindings = bindingList
+    .filter((binding) => {
+      return !binding.hide
+    })
+    .map((binding) => {
+      const flag = binding.customFlagData
+
+      return (
+        <Button
+          key={binding.id}
+          disabled={binding.id === currentBinding.id}
+          isLoading={(loading && binding.id === currentBinding.id) || isLoading}
+          onClick={() => {
+            onSelectBinding(binding)
+            setLoading(true)
+          }}
+        >
+          <LabelOption binding={binding} display={display} flag={flag} />
+        </Button>
+      )
+    })
+
   return (
-    <ul
-      hidden={!open}
-      className={`absolute z-9999 list top-1 ph0 mh0 mt5 bg-base ${handles.list}`}
-    >
-      {bindingInfo
-        .filter((binding) => {
-          return binding.id !== currentBinding.id && !binding.hide
-        })
-        .map((binding) => (
-          // eslint-disable-next-line jsx-a11y/no-noninteractive-element-interactions
-          <li
-            key={binding.id}
-            className={`${handles.listElement} t-action--small pointer f5 pa3 hover-bg-muted-5 tc`}
-            onClick={() => {
-              onSelectBinding(binding)
-            }}
-            onKeyDown={() => {
-              onSelectBinding(binding)
-            }}
-            onMouseDown={(e) => e.preventDefault()}
-          >
-            <span>{binding.label}</span>
-          </li>
-        ))}
-    </ul>
+    <div className={`${handles.listContainer}`}>
+      <ButtonGroup buttons={mappedBindings} />
+    </div>
   )
 }