+++ react/typings/bindings.d.ts
@@ -41,6 +41,7 @@ interface BindingsSaved {
   show: boolean
   translatedLocales: BindingTranslation[]
   externalRedirectData: ExternalRedirectData | null
+  customFlagData: CustomFlagData | null
 }
 
 interface UpdateSalesChannelVariables {
@@ -62,10 +63,20 @@ interface ExternalRedirectData {
   redirectUrl: boolean
 }
 
+interface CustomFlagData {
+  url: string
+  isCustom: boolean
+}
+
 interface TranslationsAndSettings extends BindingTranslation {
   externalRedirectData: ExternalRedirectData | null
+  customFlagData: CustomFlagData | null
 }
 
 interface ShowBindings {
   [key: string]: boolean
 }
+
+type FlagDisplay = 'text' | 'flag' | 'combined'
+
+type SettingType = 'externalRedirectData' | 'customFlagData'